Fix for disappearing spawns
Fix for lack of spawn delay time when searching for new pokemon. If the
TTH indicates the spawn should appear 8 minutes past the hour, the
scanner shouldn’t look until 8 minutes 10 seconds, because it takes a
few seconds for the actual appearance.
Fix for lack of missed_count reset when spawn *is* found